Add files via upload
"Interaction File for an Artistic Display on the Occasion of Mother's and Women's Day"

This file contains an interactive code to be used in a special art display on the occasion of Mother's and Women's Day. This code includes a set of interactive activities such as:

Electronic Tree of Thanks: To display words of thanks and appreciation to mothers.
The Role of Women in the World: To display information about women in various fields.
Historical Pictures of Women Through the Ages: To display pictures of creative women from different times.
Describing Mothers: The audience describes their mothers in one word to display interactions.
Inspirational Quotes: To display distinctive quotes about mothers and women.